Fix 136041 -- compability refactor about BigDecimal with JDK 1.4.2 and 1.5.

+++ model/org.eclipse.birt.report.model/src/org/eclipse/birt/report/model/api/elements/structures/PropertyBinding.java
@@ -183,7 +183,7 @@ public class PropertyBinding extends Structure
 
 	public void setID( long id )
 	{
-		this.id = new BigDecimal( id );
+		this.id = new BigDecimal( (double)id );
 	}
 
 	/**
